+++ Package.swift
@@ -26,8 +26,8 @@ let package = Package(
   targets: [
     .binaryTarget(
       name: "BrazeKit",
-      url: "https://github.com/braze-inc/braze-swift-sdk/releases/download/5.5.1/BrazeKit.zip",
-      checksum: "7e1060aec7b77c3e2c5648ec7308c1f65686a904fb971c325fc2b3d80540503b"
+      url: "https://github.com/braze-inc/braze-swift-sdk/releases/download/5.6.0/BrazeKit.zip",
+      checksum: "c86831ad7bd173f75d25833f6e6648ff3f021eebd41224f7c0c9880521ba8b0c"
     ),
     .target(
       name: "BrazeKitResources",
@@ -42,18 +42,18 @@ let package = Package(
     ),
     .binaryTarget(
       name: "BrazeLocation",
-      url: "https://github.com/braze-inc/braze-swift-sdk/releases/download/5.5.1/BrazeLocation.zip",
-      checksum: "245089b669640e15fd9a1a4e97756dd39e7223dbc67545bbb2a283a600b71ff9"
+      url: "https://github.com/braze-inc/braze-swift-sdk/releases/download/5.6.0/BrazeLocation.zip",
+      checksum: "30bc222ead44126b8f71a73de2f1cced8db57317b9a51a61f3dcb9930d31fb68"
     ),
     .binaryTarget(
       name: "BrazeNotificationService",
-      url: "https://github.com/braze-inc/braze-swift-sdk/releases/download/5.5.1/BrazeNotificationService.zip",
-      checksum: "02579b6586bf8aee37530241b195e81d616ab9c6e18a516a6ec35d34f3a36df4"
+      url: "https://github.com/braze-inc/braze-swift-sdk/releases/download/5.6.0/BrazeNotificationService.zip",
+      checksum: "26b92b5b17df840b7a0f08d2ac56d04dc4e415a16cd8da38144c0cd0600efff4"
     ),
     .binaryTarget(
       name: "BrazePushStory",
-      url: "https://github.com/braze-inc/braze-swift-sdk/releases/download/5.5.1/BrazePushStory.zip",
-      checksum: "86f6dab41fa804c06f48b5ef828a8d3eecbbeda9a8c9d077598d2b1ede79eef1"
+      url: "https://github.com/braze-inc/braze-swift-sdk/releases/download/5.6.0/BrazePushStory.zip",
+      checksum: "f7af4633a4ebd036c7beeecc84ef9289b6e393c590ea726d15aa6359befa71e4"
     ),
   ]
 )