+++ Sources/BrazeUI/ContentCardUI/Cells/ContentCardUIClassicCell.swift
@@ -22,7 +22,7 @@ extension BrazeContentCardUI {
       container.addSubview(textStack)
 
       installInternalConstraints()
-      applyAttributes()
+      applyAttributes(attributes)
     }
 
     /// Does not support interface-builder / storyboards.
@@ -58,8 +58,8 @@ extension BrazeContentCardUI {
       highlightable = card.clickAction != .none
     }
 
-    open override func applyAttributes() {
-      super.applyAttributes()
+    open override func applyAttributes(_ attributes: Attributes) {
+      super.applyAttributes(attributes)
 
       let padding = attributes.padding
       textStackContraints[0].constant = padding.left