+++ Sources/BrazeUI/ContentCardUI/Cells/ContentCardUIClassicImageCell.swift
@@ -39,7 +39,7 @@ extension BrazeContentCardUI {
       container.addSubview(cardStack)
 
       installInternalConstraints()
-      applyAttributes()
+      applyAttributes(attributes)
     }
 
     /// Does not support interface-builder / storyboards.
@@ -84,8 +84,8 @@ extension BrazeContentCardUI {
       highlightable = card.clickAction != .none
     }
 
-    open override func applyAttributes() {
-      super.applyAttributes()
+    open override func applyAttributes(_ attributes: Attributes) {
+      super.applyAttributes(attributes)
 
       let padding = attributes.padding
       cardStackConstraints[0].constant = padding.left